rename mysql db config file.
