+++ api/migrate_db.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 
 from api.models.task import Base
-from api.config.config import DBConfig
+from api.config.mysql_config import DBConfig
 
 DB_URL="mysql+pymysql://" + DBConfig().__str__()
 engine = create_engine(DB_URL, echo=True)
